% Auto-generated: do not edit by hand
\name{chat}

\alias{chat}

\title{Chat component}

\description{
Main chat container
}

\usage{
chat(children=NULL, id=NULL, className=NULL, height=NULL,
style=NULL, width=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. Content of the chat}

\item{id}{Character. Unique ID to identify this component in Dash callbacks.}

\item{className}{Character. Often used with CSS to style elements with common properties.}

\item{height}{Numeric. Height of the chat container}

\item{style}{Logical | numeric | character | named list | unnamed list. Defines CSS styles which will override styles previously set.}

\item{width}{Numeric. Width of the chat container}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

