% Auto-generated: do not edit by hand
\name{chatMessage}

\alias{chatMessage}

\title{ChatMessage component}

\description{
Display single chat message
}

\usage{
chatMessage(children=NULL, id=NULL, avatar=NULL, className=NULL,
direction=NULL, style=NULL, timestamp=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. Content of the message}

\item{id}{Character. Unique ID to identify this component in Dash callbacks.}

\item{avatar}{Character. Avatar}

\item{className}{Character. Often used with CSS to style elements with common properties.}

\item{direction}{Character. Direction of the mssage: received or outgoing}

\item{style}{Logical | numeric | character | named list | unnamed list. Defines CSS styles which will override styles previously set.}

\item{timestamp}{Numeric. Time}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

