% Auto-generated: do not edit by hand
\name{chatMessageList}

\alias{chatMessageList}

\title{ChatMessageList component}

\description{
Display list of chat messages
}

\usage{
chatMessageList(children=NULL, id=NULL, avatarOutgoing=NULL,
avatarReceived=NULL, className=NULL, style=NULL,
timestamp=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. ChatMessageList content}

\item{id}{Character. Unique ID to identify this component in Dash callbacks.}

\item{avatarOutgoing}{Logical. Indicates whether the outgoing message avatar should be displayed.}

\item{avatarReceived}{Logical. Indicates whether the recieved message avatar should be displayed.}

\item{className}{Character. Often used with CSS to style elements with common properties.}

\item{style}{Logical | numeric | character | named list | unnamed list. Defines CSS styles which will override styles previously set.}

\item{timestamp}{Logical. Indicates whether the message timestamp should be displayed.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

