% Auto-generated: do not edit by hand
\name{chatMessageTyping}

\alias{chatMessageTyping}

\title{ChatMessageTyping component}

\description{
Display user typing message
}

\usage{
chatMessageTyping(id=NULL, avatar=NULL, className=NULL, direction=NULL,
style=NULL)
}

\arguments{
\item{id}{Character. Unique ID to identify this component in Dash callbacks.}

\item{avatar}{Character. Avatar}

\item{className}{Character. Often used with CSS to style elements with common properties.}

\item{direction}{Character. Direction of the mssage: received or outgoing}

\item{style}{Logical | numeric | character | named list | unnamed list. Defines CSS styles which will override styles previously set.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

